Guard against emitted invalid undefined range
