+++ src/SIL.XForge.Scripture/ClientApp/src/app/serval-administration/date-range-picker.component.html
@@ -10,7 +10,7 @@
       matSuffix
       type="button"
       (click)="resetToDefaultDateRange()"
-      id="reset-button"
+      data-test-id="reset-button"
       aria-label="Reset date range"
     >
       <mat-icon>close</mat-icon>
