+++ src/SIL.XForge.Scripture/ClientApp/src/app/serval-administration/serval-builds.component.ts
@@ -276,6 +276,7 @@ export class ServalBuildsComponent extends DataLoadingComponent implements OnIni
   }
 
   protected onDateRangeChange(range: NormalizedDateRange | undefined): void {
+    if (range == null) return;
     this.dateRange$.next(range);
   }
